@original-brownbear 2017-09-12
@srowen thanks, all comments addressed I think :)