@original-brownbear 2017-09-13
@srowen I think we have that already here https://github.com/apache/spark/blob/master/common/unsafe/src/test/java/org/apache/spark/unsafe/types/UTF8StringSuite.java#L103

`("你好123".getBytes(StandardCharsets.UTF_8)[0] & 0xFF) == 0b11100100` :)