@srowen 2017-09-14
CC @davies 
By the way do you have any measurements that show the speed up? I imagine it's faster, but mostly curious if it's still a win for short strings.