@srowen 2017-09-14
Yeah it would be great if it were strictly faster. For < 7 bytes, the new loop doesn't execute, so it doesn't matter, and it just proceeds with the old loop, which should be the same. The only overhead I can see is getting the offsets. What if you just short-circuited right after computing wordMax and check `wordMax > 0`? then I don't see how it would be slower. It'd be simpler too.