@original-brownbear 2017-09-14
@srowen updated, we don't even need to break out via `wordMax > 0`, we can just get the offsets from `other` and save one level of indirect in the second loop by using `(Platform.getByte(rbase, roffset + i) & 0xFF)` :)